refactor legend config
